AnomalyDetectionService/anomaly_detection/requirements.txt

Flask==2.3.2
flask-restful==0.3.9
gunicorn==21.2.0
joblib==1.2.0
kafka-python==2.0.2
matplotlib==3.7.1
numpy==1.26.0
pandas==2.1.1
requests==2.31.0
scikit-learn==1.3.0
tensorflow==2.15.0
torchvision==0.16.0
